refactor: input파일 분리
squash
